Encode/decode protobufs as base64
The utf-8 encoding doesn't work for arbitrary bytes.
The problem was the same as described in https://github.com/mymarilyn/clickhouse-driver/issues/25
but since `RunMeta´ bytes are not fixed-size, the solutions there don't apply.
By first base64-encoding the bytes it is ensured that they can be decoded as a string without problems.