[NumPyBackend] Prevent ``preallocate=0`` from creating object arrays.
`grow_append` cannot know if ``preallocate = 0`` was used: it only
looks at the `rigid` value to determine how to append.
Because of this, will always fail when we use `preallocate = 0` with
tensor variables, since then the shapes of `target` and `extension`
don’t match.
A simple fix is to simply deactivate the special behavior for
`preallocate = 0`.

- This commit extends `test_growing` with a `preallocate` parameter,
  so that we test both cases where it is 0 and positive.
- We also fix `test_growing` to match the new behavior of `grow_append`
  introduced in #ea812b0, where data arrays always grow by at least 10.